requirements.txt
streamlit
matplotlib.pyplot
seaborn
pandas
numpy
requests
urllib.request
json
plotly
plotly.express
requests